Move reading env variables to the heir of Flask class
