+++ fb_bot.py
@@ -1,12 +1,18 @@
 from datetime import datetime
+from enum import Enum
+import os
 
+from dotenv import load_dotenv
 from flask import Flask, request
 from flask_apscheduler import APScheduler
+from redis import Redis
 
-from helpers.fb_envvar_handlers import get_env_vars
-from helpers.fb_action_handlers import handle_menu_caching, handle_user_input
-
-AUTH, DB, FB_TOKEN, VERIFY_TOKEN, state = get_env_vars()
+from helpers.fb_action_handlers import (
+    handle_menu_caching,
+    handle_user_input,
+    State,
+)
+from helpers.token_handers import AuthToken
 
 
 class Config:
@@ -14,7 +20,24 @@ class Config:
     SCHEDULER_TIMEZONE = "Europe/Paris"
 
 
-app = Flask(__name__)
+class MyFlaskApp(Flask):
+    load_dotenv()
+
+    client_id = os.getenv('CLIENT_ID')
+    client_secret = os.getenv('CLIENT_SECRET')
+    auth = AuthToken(client_id, client_secret)
+
+    redis_password = os.getenv('REDIS_PASSWORD', '')
+    redis_host = os.getenv('REDIS_HOST', 'localhost')
+    redis_port = os.getenv('REDIS_PORT', 6379)
+    db = Redis(host=redis_host, port=redis_port, password=redis_password)
+
+    fb_token = os.getenv('PAGE_ACCESS_TOKEN')
+    verify_token = os.getenv('VERIFY_TOKEN')
+    state = State.MENU
+
+
+app = MyFlaskApp(__name__)
 app.config.from_object(Config())
 
 scheduler = APScheduler()
@@ -24,8 +47,7 @@ scheduler.start()
 
 @scheduler.task('interval', minutes=1, next_run_time=datetime.now())
 def cache_menu():
-    handle_menu_caching(AUTH, DB)
-    print('cached')
+    handle_menu_caching(app.auth, app.db)
 
 
 @app.route('/', methods=['GET'])
@@ -34,7 +56,7 @@ def verify():
 
     args = request.args
     if args.get('hub.mode') == 'subscribe' and args.get('hub.challenge'):
-        if not request.args.get('hub.verify_token') == VERIFY_TOKEN:
+        if not request.args.get('hub.verify_token') == app.verify_token:
             return 'Verification token mismatch', 403
         return request.args['hub.challenge'], 200
 
@@ -59,8 +81,14 @@ def webhook():
                 if message or postback:
                     message = message['text'] if message else None
                     payload = postback['payload'] if postback else None
-                    state = handle_user_input(
-                        DB, state, user_id, AUTH, FB_TOKEN, message, payload
+                    app.state = handle_user_input(
+                        db=app.db,
+                        state=app.state,
+                        user_id=user_id,
+                        auth=app.auth,
+                        fb_token=app.fb_token,
+                        message=message,
+                        payload=payload,
                     )
 
     return "ok", 200
